Initial port of OpenAI samples into tests
The samples have been set up with assertions to check that OpenAI
services do not have issues in Android. The test class is similar to the
existing OpenAIAsyncClientTest albeit greatly simplified with JUnit 4 annotations to run in
Android. The printout statements have temporarily been changed to Android logs until the samples are confirmed
to be fully functional. Afterwards, they could be removed.